rename file 'users' to 'userController'
